requirements.txt
absl-py==0.13.0
accelerate==0.16.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.3.0
analytics-python==1.4.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
backoff==1.10.0
bcrypt==3.2.2
beartype==0.12.0
beautifulsoup4==4.11.1
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.6
configparser==5.0.2
cryptography==37.0.2
cycler==0.10.0
datasets==2.9.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.6.0
einops-exts==0.0.4
ema-pytorch==0.1.4
entrypoints==0.3
fastapi==0.78.0
ffmpy==0.3.0
filelock==3.0.12
FrEIA @ git+https://github.com/VLL-HD/FrEIA.git@1779d1fba1e21000fda1927b59eeac0a6fcaa284
frozenlist==1.3.0
fsspec==2022.5.0
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.5
googledrivedownloader==0.4
gql==0.2.0
gradio==3.0.17
graphql-core==1.1
grpcio==1.39.0
h11==0.13.0
huggingface-hub==0.12.0
hydra-core==1.2.0
idna==3.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.8.1
importlib-resources==5.7.1
intensity-normalization==2.1.4
ipykernel==6.2.0
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
kornia==0.6.5
linkify-it-py==1.0.3
llvmlite==0.33.0
lmdb==1.4.0
Markdown==3.3.4
markdown-it-py==2.1.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mdit-py-plugins==0.3.0
mdurl==0.1.1
mistune==2.0.4
monotonic==1.6
multidict==6.0.2
multiprocess==0.70.14
munch==2.5.0
nbclient==0.5.4
nbconvert==7.2.9
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nibabel==3.2.2
notebook==6.4.3
notebook-as-pdf==0.5.0
numba==0.50.1
numpy==1.21.6
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
omegaconf==2.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
orjson==3.7.2
packaging==22.0
pandas==1.3.2
pandocfilters==1.4.3
paramiko==2.11.0
parso==0.8.2
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plyfile==0.7.4
pretrainedmodels==0.7.4
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
ptflops==0.6.6
ptyprocess==0.7.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.14.1
pydantic==1.9.1
pyDeprecate==0.3.2
pydub==0.25.1
pyee==7.0.4
Pygments==2.10.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyppeteer==0.2.2
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-louvain==0.15
python-multipart==0.0.5
pytorch-lightning==1.6.4
pytorch-metric-learning==0.9.99
pytorch-warmup==0.1.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==305
pywinpty==2.0.10
PyYAML==5.4.1
pyzmq==22.2.1
qudida==0.0.4
rdflib==6.0.0
regex==2022.10.31
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.18.0
rsa==4.7.2
scikit-fuzzy==0.4.2
scikit-image==0.19.3
scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.7.1
Send2Trash==1.8.0
sentry-sdk==1.3.1
setproctitle==1.2.3
Shapely==1.8.2
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
statsmodels==0.13.5
subprocess32==3.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.5.1
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.11.2
timm==0.5.4
tinycss2==1.2.1
tokenizers==0.13.2
torch==1.12.1+cu116
torch-geometric==1.7.2
torchaudio==0.12.1+cu116
torchfile==0.1.0
torchmetrics==0.9.0
torchnet==0.0.4
torchvision==0.13.1+cu116
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
transformers==4.26.0
typing_extensions==4.2.0
uc-micro-py==1.0.1
urllib3==1.26.6
uvicorn==0.17.6
visdom==0.1.8.9
wandb==0.12.17
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
xxhash==3.2.0
yarl==1.7.2
zipp==3.5.0
